use typescript for AddConnections
